cuda monitor refactored, part 3
+ statistics refactored, part 4
+ Incubator example updated
+ /MP compiler flag for MSVC

tab widget for statistics window

statistics refactored, part 5

statistics refactored, part 6

show simple histogram with hard-coded max age

take max age into account for age histogram

format

draw grid lines in histogram

Update README.md

Update README.md

+ timeout bug in statistics kernel fixed
+ jungle example finished
+ show max value in histogram

Update README.md

Update README.md

Update README.md

swarm examples

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

delete only cell connections (not cell itself) when forces exceed threshold

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Stormy Night demo

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

Update README.md

gcc compile errors addressed

Stormy Night updated

Update README.md

tests repaired

Update README.md

thin out timeline statistics when too large

statistics refactored, part 7

+++ source/EngineTests/NeuronTests.cpp
@@ -16,7 +16,7 @@ public:
     ~NeuronTests() = default;
 
 protected:
-    float scaledSigmoid(float value) const { return 2.0f / (1.0f + std::expf(-value)) - 1.0f; }
+    float scaledSigmoid(float value) const { return 2.0f / (1.0f + std::exp(-value)) - 1.0f; }
 };
 
 TEST_F(NeuronTests, bias)
